#1991 Shadow copy service

+++ src/main/java/eu/dzhw/fdz/metadatamanagement/questionmanagement/service/QuestionService.java
@@ -4,8 +4,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import eu.dzhw.fdz.metadatamanagement.common.domain.ShadowCopyDeleteNotAllowedException;
+import eu.dzhw.fdz.metadatamanagement.common.service.ShadowCopyService;
+import eu.dzhw.fdz.metadatamanagement.projectmanagement.domain.ProjectReleasedEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.context.event.EventListener;
 import org.springframework.data.rest.core.annotation.HandleAfterCreate;
 import org.springframework.data.rest.core.annotation.HandleAfterDelete;
 import org.springframework.data.rest.core.annotation.HandleAfterSave;
@@ -60,6 +64,12 @@ public class QuestionService {
   @Autowired
   private RelatedPublicationChangesProvider relatedPublicationChangesProvider;
 
+  @Autowired
+  private ShadowCopyService<Question> shadowCopyService;
+
+  @Autowired
+  private QuestionShadowCopyDataSource questionShadowCopyDataSource;
+
   /**
    * Delete all questions when the dataAcquisitionProject was deleted.
    *
@@ -92,6 +102,9 @@ public class QuestionService {
     try (Stream<Question> questions =
         questionRepository.streamByDataAcquisitionProjectId(dataAcquisitionProjectId)) {
       questions.forEach(question -> {
+        if (question.isShadow()) {
+          throw new ShadowCopyDeleteNotAllowedException();
+        }
         eventPublisher.publishEvent(new BeforeDeleteEvent(question));
         questionRepository.delete(question);
         eventPublisher.publishEvent(new AfterDeleteEvent(question));
@@ -199,5 +212,19 @@ public class QuestionService {
     elasticsearchUpdateQueueService.enqueueUpsertsAsync(
         () -> questionRepository.streamIdsByIdIn(questionIds),
         ElasticsearchType.questions);
+
+    elasticsearchUpdateQueueService.enqueueUpsertsAsync(
+        () -> questionRepository.streamIdsByMasterIdInAndShadowIsTrueAndSuccessorIdIsNull(
+            questionIds), ElasticsearchType.questions);
+  }
+
+  /**
+   * Create shadow copies for questions on project release.
+   * @param projectReleasedEvent Released project event
+   */
+  @EventListener
+  public void onProjectReleaseEvent(ProjectReleasedEvent projectReleasedEvent) {
+    shadowCopyService.createShadowCopies(projectReleasedEvent.getDataAcquisitionProject(),
+        projectReleasedEvent.getPreviousReleaseVersion(), questionShadowCopyDataSource);
   }
 }
